Fix conversion warnings for ints
Numbers should not be automatically converted to a less precise representation.
Sizes of vectors are unsigned long and this should be used to work with them.
long and short is more precise than int, please decide the precision. int will
leave the decision to the architecture. int is ok for the root output tree.